V1.0.5 Removed all uses of xargs. Fixed all skoonieini file abs paths Updated help output
All uses of xargs were removed because it was having issues when the lines
read in from the interface skoonieini file contained apostrophes. Uses of
xargs were replaced with pure bash.

Most of the skoonieini file absolute paths beings set were missing a "/".

Updated help output to fix the example usages of addInterface

Signed-off-by: Hunter Schoonover <hunter@skoonie.com>