WireGuard Interfaces created by wg-skoonie now allow comms between clients
PostUp and PostDown commands are now added to each WireGuard interface
server configuration file generated.

PostUp adds rules to iptables that allow client devices on the VPN to
communicate with other client devices on the VPN so long as both devices
are within the same subnet.

PostDown removes the rules from iptables when the WireGuard interfaces is
brought down.

Signed-off-by: Hunter Schoonover <hunter@skoonie.com>