Added cronjob connectivity checker that can restart WG interface .
This was added after an extended power outage caused the server's dynamic
IP address to change. DuckDNS was being used to provide Dynamic DNS.

While the new IP address was provided to DuckDNS, the client devices never
acknowledged the change in IP address. This is because the WireGuard only
resolves the server address when it firsts starts up.

To prevent this issue in the future, cronjob scripts that ping the server's
IP address on the VPN subnet are now generated for each supported OS
(except Windows). When the server can't be reached, the WireGuard
interface is brought down, a call to "getent ahosts" is performed to force
device DNC cache refresh, and then the WireGuard interface is brought back
up.

The setup scripts now automatically install and set up the cronjob
connectivity checker script to run every 15 minutes on the client device.

Signed-off-by: Hunter Schoonover <hunter@skoonie.com>