Code refactoring so that generated setup scripts share more code.
There was a lot of repeated code between the functions generating the
setup scripts for devices running traditional ifupdown or netplan. This
was wasted space, difficult to maintain, and prone to bugs.

The code was refactored so that as much code as possible is shared. This
may have even fixed unknown bugs.

Signed-off-by: Hunter Schoonover <hunter@skoonie.com>