WireGuard config is now saved immediately after a device is added.
Prior to this, wg-skoonie relied on WireGuard to save its own config, which
WireGuard does successfully if the interface is brought down properly.
However, if the interface is halted unexpectedly (such as power outage),
the config is not saved and the devices in WireGuard that were not yet
saved to file are lost.

Signed-off-by: Hunter Schoonover <hunter@skoonie.com>