WireGuard config is now saved immediately after a device is removed.
Prior to this, wg-skoonie relied on WireGuard to save its own config, which
WireGuard does successfully if the interface is brought down properly.
However, if the interface is halted unexpectedly (such as power outage),
the config is not saved properly.

In such a case, a device removed could have been added back in when the
device starts up again.

Signed-off-by: Hunter Schoonover <hunter@skoonie.com>